# [ez][HUD] Optimize query for TTRS, previous commits on PR