Small optimizations for some queries, not super necessary 

TTRS:
The query_logs table claims it takes ~17000ms on avg, but in the console it usually takes ~5000ms.  Using lookup brings to about ~1000ms on the console

PR commits (used on HUD PR page to fetch previous commits)
This was already pretty fast, but its one of the most used queries.  Original time: ~1500ms.  Moving the joins brings it to ~100ms